Implement "extra text" RAW spans.
Allow adding EXTRA_TEXT type elements to RAW
span lists. These "add" strings of text to the
middle of the parsed data without modifying the
underlying charbuf; the parser reads their text
whenever it comes by them but they are not taken
into account when calculating correct offsets for
matched text. This is for simulating situations
in peg-markdown where extra strings (like "\n\n")
are added to the middle of the RAW data. The
highlighting has broken; need further fixes to
get back on track.