Add example Cocoa project + helper class.
The class HGMarkdownHighlighter is supposed to
implement the task of highlighting any NSTextView
automagically. It works ok right now, running the
parser in a background thread (one at a time),
keeping the resulting element arrays in memory and
applying the highlighting only to the visible area
(also whenever the text view scrolls). There are
some weird small bugs (jerkiness in scrolling)
that this causes, though, and which I have to
investigate. Also, scrolling a long document uses
*waay too much CPU* because each time we have to
go through all of the elements in the array of
linked lists and see if they fall into the visible
range of the text, and then apply the
highlighting. If this is really the approach we
want to take, we need to index the element array
by location/offset.