Revert 'references on last line' fix.
Referring to changeset 
603f94653641facd42fc234b274dcb5217398903.
The fix was erroneous and caused a regression
where titles that contained their delimiter
characters were not matched (e.g. "title "with"
double quotes inside") even though they should be.
This change fixes this regression but the original
problem (titles not matched if the reference
line ends in EOF) remains.