Highlight links only if their reference exists.
We do this by performing a first pass over the
document where we only gather references (just
like peg-markdown does). Then we store the linked
list of reference elements in the 'references'
member of the parsing data struct. We store the
label of each reference in the element and can
then use that to check for each link whether the
reference for its label exists.