Fix crash if an element type name was not defined.
The element type name -related functions in the
style parser expected all element types defined
in the element_type enum to have a name defined
in the elem_type_names array in
get_element_type_names(). If someone adds new
element types to markdown_definitions.h but does
not define names for them in styleparser.leg, this
caused the style parser to crash. Now we don't
expect all element_type values to have a name
defined in that array.

Thanks to Fletcher T. Penney for noticing this.