Cocoa: Fix crash caused by a race condition.
An interleaving of threads could exist where
-threadDidExit: (which runs in the background
thread) and -requestParsing (which runs in the
main thread) might both try to free()
currentHighlightText, resulting in a crash. We
(hopefully!) fix this by setting self.workerThread
to nil *after* free()ing currentHighlightText
in -threadDidExit: -- this will keep the body of
-requestParsing from executing before
-threadDidExit: has had the chance to set the
currentHighlightText pointer to NULL after
free()ing it.

Thanks to Fletcher Penney for alerting me to this
issue.