Fix memory leaks in parser.
Free `text`, `label` and `address` members of
element structures in free_elements(). For this
to work we need to make sure that elements own
any non-NULL strings they have behind these
pointers. Also use a separate offset integer for
traversing the `text` value in EXTRA_TEXT
elements (instead of moving the pointer itself,
which would leave it unaligned, causing a crash
when we try to free it).