Fix handing of unicode in links.
We got the values for the `address` and `label`
members of pmh_elements by copying `yytext` in
the parser's grammar actions, which meant that
we'd have versions of those strings *without* any
UTF-8 continuation bytes, as the parser strips
them out before starting the parsing. So we need
to keep track of each byte we strip from the
original input so that we can map offsets in our
internal, modified buffer to the corresponding
offsets in the original input. Then we can copy
the address and label values from the original
buffer and have the unmodified UTF-8 versions of
them.