Fix character index bug in Cocoa highlighting.
The parser returns us indexes/offsets for unicode
code points, but the NSMutableAttributedStrings
we highlight are indexed by UTF-16 units. A UTF-16
unit is often only one code point, but for example
for some Chinese characters (that would be encoded
in UTF-8 as four bytes) a surrogate pair of two
UTF-16 units is needed. If the highlighted
document contained such characters, our indexes
would be incorrect and thus the wrong parts of the
output would be highlighted.

We fix this by manually converting each code point
range to a character range after running the
parser.

Thanks to Chen Luo for reporting the bug.