Cocoa: Make our timer ignore OS event tracking.
Since we were scheduling our highlighting timer
in the run loop with only the default mode, when
e.g. the mouse key was depressed it would not
fire until the key was released (when the key is
depressed the run loop is in "event tracking"
mode). We fix this by scheduling the timer in
"all common modes".