Cocoa: Disable some problematic text view features
Automatic text replacement in particular causes
our highlighting to either get messed up (if the
automatic highlighting when the text changes is
disabled) or look a bit ugly (if the automatic
highlighting is on, the highlighted sections will
"jump" sometimes when automatic text replacement
does something). To make the demo app work a bit
better, we disable this and some other "fancy"
features in the text views.